// Acoes/src/interfaces/UsuarioUpdateData.ts
export interface UsuarioUpdateData {
    nome: string;
    email: string;
    telefone: string;
    senha: string;
    cpf: string;
    dataNascimento: string;
    genero: number;
    cidade: string;
    estado: string;
    caminhoCarteira: string;
    caminhoCertidao: string;
    whatsapp: string;
    servicos: Array<number>;
    valorMax: number;
    valorMin: number;
}